requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
scikit-learn==1.3.2
